+++ app/src/main/java/com/team/bottles/di/usecase/ProfileUseCaseModule.kt
@@ -1,9 +1,5 @@
 package com.team.bottles.di.usecase
 
-import com.team.bottles.core.domain.auth.usecase.LoginWithKakaoUseCase
-import com.team.bottles.core.domain.auth.usecase.LoginWithKakaoUseCaseImpl
-import com.team.bottles.core.domain.auth.usecase.WebViewConnectUseCase
-import com.team.bottles.core.domain.auth.usecase.WebViewConnectUseCaseImpl
 import com.team.bottles.core.domain.profile.usecase.CreateIntroductionUseCase
 import com.team.bottles.core.domain.profile.usecase.CreateIntroductionUseCaseImpl
 import com.team.bottles.core.domain.profile.usecase.GetUserProfileUseCase
@@ -17,17 +13,7 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class UseCaseModule {
-
-    @Binds
-    abstract fun bindsLoginWithKaKaoUseCase(
-        useCaseImpl: LoginWithKakaoUseCaseImpl
-    ): LoginWithKakaoUseCase
-
-    @Binds
-    abstract fun bindsWebViewConnectUseCase(
-        useCaseImpl: WebViewConnectUseCaseImpl
-    ): WebViewConnectUseCase
+abstract class ProfileUseCaseModule {
 
     @Binds
     abstract fun bindsCreateIntroductionUseCase(
